utils.isjQuery: Improve `Identifier` detection
Previously testing `$div` in `foo($div)` would return `false`
as it assumed that if `$div`'s parent node was a `CallExpression`,
that `$div` was the node being invoked, and so only returned
`true` iff the jQuery constructor was detected.

Add in a `node.parent.calleee = node` check to fix this.

Add some extra unit tests to variable-pattern, although none
of these excercise the new check.